fixup(EVDT-007): split pure logic into upsert-rules.ts so tests don't need DATABASE_URL
CI test step doesn't have DATABASE_URL set; importing upsert.ts pulls in
@/db/client which throws at module load. Moving the pure decision logic
to upsert-rules.ts (no db import) means the unit test only loads the
pure code path.

Test runs CI-clean now; behavior unchanged.
